+++ tests/CMYKATest.php
@@ -19,11 +19,11 @@ class CMYKATest extends TestCase
 {
     public function testInterface()
     {
-        $cmyk = new CMYKA(
-            $cyan = new Cyan(10),
-            $magenta = new Magenta(20),
-            $yellow = new Yellow(30),
-            $black = new Black(40),
+        $cmyk = CMYKA::from(
+            $cyan = Cyan::at(10),
+            $magenta = Magenta::at(20),
+            $yellow = Yellow::at(30),
+            $black = Black::at(40),
         );
 
         $this->assertSame($cyan, $cmyk->cyan());
@@ -33,12 +33,12 @@ class CMYKATest extends TestCase
         $this->assertSame(1.0, $cmyk->alpha()->toFloat());
         $this->assertSame('device-cmyk(10%, 20%, 30%, 40%)', $cmyk->toString());
 
-        $cmyka = new CMYKA(
-            new Cyan(10),
-            new Magenta(20),
-            new Yellow(30),
-            new Black(40),
-            $alpha = new Alpha(0.5),
+        $cmyka = CMYKA::from(
+            Cyan::at(10),
+            Magenta::at(20),
+            Yellow::at(30),
+            Black::at(40),
+            $alpha = Alpha::of(0.5)->unwrap(),
         );
 
         $this->assertSame($alpha, $cmyka->alpha());
@@ -47,14 +47,14 @@ class CMYKATest extends TestCase
 
     public function testAddCyan()
     {
-        $cmyk = new CMYKA(
-            new Cyan(10),
-            new Magenta(20),
-            new Yellow(30),
-            new Black(40),
+        $cmyk = CMYKA::from(
+            Cyan::at(10),
+            Magenta::at(20),
+            Yellow::at(30),
+            Black::at(40),
         );
 
-        $cmyk2 = $cmyk->addCyan(new Cyan(1));
+        $cmyk2 = $cmyk->addCyan(Cyan::at(1));
 
         $this->assertInstanceOf(CMYKA::class, $cmyk2);
         $this->assertNotSame($cmyk, $cmyk2);
@@ -72,14 +72,14 @@ class CMYKATest extends TestCase
 
     public function testSubtractCyan()
     {
-        $cmyk = new CMYKA(
-            new Cyan(10),
-            new Magenta(20),
-            new Yellow(30),
-            new Black(40),
+        $cmyk = CMYKA::from(
+            Cyan::at(10),
+            Magenta::at(20),
+            Yellow::at(30),
+            Black::at(40),
         );
 
-        $cmyk2 = $cmyk->subtractCyan(new Cyan(1));
+        $cmyk2 = $cmyk->subtractCyan(Cyan::at(1));
 
         $this->assertInstanceOf(CMYKA::class, $cmyk2);
         $this->assertNotSame($cmyk, $cmyk2);
@@ -97,14 +97,14 @@ class CMYKATest extends TestCase
 
     public function testAddMagenta()
     {
-        $cmyk = new CMYKA(
-            new Cyan(10),
-            new Magenta(20),
-            new Yellow(30),
-            new Black(40),
+        $cmyk = CMYKA::from(
+            Cyan::at(10),
+            Magenta::at(20),
+            Yellow::at(30),
+            Black::at(40),
         );
 
-        $cmyk2 = $cmyk->addMagenta(new Magenta(1));
+        $cmyk2 = $cmyk->addMagenta(Magenta::at(1));
 
         $this->assertInstanceOf(CMYKA::class, $cmyk2);
         $this->assertNotSame($cmyk, $cmyk2);
@@ -122,14 +122,14 @@ class CMYKATest extends TestCase
 
     public function testSubtractMagenta()
     {
-        $cmyk = new CMYKA(
-            new Cyan(10),
-            new Magenta(20),
-            new Yellow(30),
-            new Black(40),
+        $cmyk = CMYKA::from(
+            Cyan::at(10),
+            Magenta::at(20),
+            Yellow::at(30),
+            Black::at(40),
         );
 
-        $cmyk2 = $cmyk->subtractMagenta(new Magenta(1));
+        $cmyk2 = $cmyk->subtractMagenta(Magenta::at(1));
 
         $this->assertInstanceOf(CMYKA::class, $cmyk2);
         $this->assertNotSame($cmyk, $cmyk2);
@@ -147,14 +147,14 @@ class CMYKATest extends TestCase
 
     public function testAddYellow()
     {
-        $cmyk = new CMYKA(
-            new Cyan(10),
-            new Magenta(20),
-            new Yellow(30),
-            new Black(40),
+        $cmyk = CMYKA::from(
+            Cyan::at(10),
+            Magenta::at(20),
+            Yellow::at(30),
+            Black::at(40),
         );
 
-        $cmyk2 = $cmyk->addYellow(new Yellow(1));
+        $cmyk2 = $cmyk->addYellow(Yellow::at(1));
 
         $this->assertInstanceOf(CMYKA::class, $cmyk2);
         $this->assertNotSame($cmyk, $cmyk2);
@@ -172,14 +172,14 @@ class CMYKATest extends TestCase
 
     public function testSubtractYellow()
     {
-        $cmyk = new CMYKA(
-            new Cyan(10),
-            new Magenta(20),
-            new Yellow(30),
-            new Black(40),
+        $cmyk = CMYKA::from(
+            Cyan::at(10),
+            Magenta::at(20),
+            Yellow::at(30),
+            Black::at(40),
         );
 
-        $cmyk2 = $cmyk->subtractYellow(new Yellow(1));
+        $cmyk2 = $cmyk->subtractYellow(Yellow::at(1));
 
         $this->assertInstanceOf(CMYKA::class, $cmyk2);
         $this->assertNotSame($cmyk, $cmyk2);
@@ -197,14 +197,14 @@ class CMYKATest extends TestCase
 
     public function testAddBlack()
     {
-        $cmyk = new CMYKA(
-            new Cyan(10),
-            new Magenta(20),
-            new Yellow(30),
-            new Black(40),
+        $cmyk = CMYKA::from(
+            Cyan::at(10),
+            Magenta::at(20),
+            Yellow::at(30),
+            Black::at(40),
         );
 
-        $cmyk2 = $cmyk->addBlack(new Black(1));
+        $cmyk2 = $cmyk->addBlack(Black::at(1));
 
         $this->assertInstanceOf(CMYKA::class, $cmyk2);
         $this->assertNotSame($cmyk, $cmyk2);
@@ -222,14 +222,14 @@ class CMYKATest extends TestCase
 
     public function testSubtractBlack()
     {
-        $cmyk = new CMYKA(
-            new Cyan(10),
-            new Magenta(20),
-            new Yellow(30),
-            new Black(40),
+        $cmyk = CMYKA::from(
+            Cyan::at(10),
+            Magenta::at(20),
+            Yellow::at(30),
+            Black::at(40),
         );
 
-        $cmyk2 = $cmyk->subtractBlack(new Black(1));
+        $cmyk2 = $cmyk->subtractBlack(Black::at(1));
 
         $this->assertInstanceOf(CMYKA::class, $cmyk2);
         $this->assertNotSame($cmyk, $cmyk2);
@@ -247,15 +247,15 @@ class CMYKATest extends TestCase
 
     public function testAddAlpha()
     {
-        $cmyk = new CMYKA(
-            new Cyan(10),
-            new Magenta(20),
-            new Yellow(30),
-            new Black(40),
-            new Alpha(0.1),
+        $cmyk = CMYKA::from(
+            Cyan::at(10),
+            Magenta::at(20),
+            Yellow::at(30),
+            Black::at(40),
+            Alpha::of(0.1)->unwrap(),
         );
 
-        $cmyk2 = $cmyk->addAlpha(new Alpha(0.1));
+        $cmyk2 = $cmyk->addAlpha(Alpha::of(0.1)->unwrap());
 
         $this->assertInstanceOf(CMYKA::class, $cmyk2);
         $this->assertNotSame($cmyk, $cmyk2);
@@ -273,14 +273,14 @@ class CMYKATest extends TestCase
 
     public function testSubtractAlpha()
     {
-        $cmyk = new CMYKA(
-            new Cyan(10),
-            new Magenta(20),
-            new Yellow(30),
-            new Black(40),
+        $cmyk = CMYKA::from(
+            Cyan::at(10),
+            Magenta::at(20),
+            Yellow::at(30),
+            Black::at(40),
         );
 
-        $cmyk2 = $cmyk->subtractAlpha(new Alpha(0.9));
+        $cmyk2 = $cmyk->subtractAlpha(Alpha::of(0.9)->unwrap());
 
         $this->assertInstanceOf(CMYKA::class, $cmyk2);
         $this->assertNotSame($cmyk, $cmyk2);