+++ tests/CyanTest.php
@@ -13,7 +13,7 @@ class CyanTest extends TestCase
 {
     public function testInterface()
     {
-        $cyan = new Cyan(50);
+        $cyan = Cyan::at(50);
 
         $this->assertSame(50, $cyan->toInt());
         $this->assertSame('50', $cyan->toString());
@@ -24,7 +24,7 @@ class CyanTest extends TestCase
         $this->expectException(InvalidValueRangeException::class);
         $this->expectExceptionMessage('-1');
 
-        new Cyan(-1);
+        Cyan::of(-1)->unwrap();
     }
 
     public function testThrowWhenValueTooHigh()
@@ -32,14 +32,14 @@ class CyanTest extends TestCase
         $this->expectException(InvalidValueRangeException::class);
         $this->expectExceptionMessage('101');
 
-        new Cyan(101);
+        Cyan::of(101)->unwrap();
     }
 
     public function testAdd()
     {
-        $cyan = new Cyan(50);
+        $cyan = Cyan::at(50);
 
-        $cyan2 = $cyan->add(new Cyan(25));
+        $cyan2 = $cyan->add(Cyan::at(25));
 
         $this->assertInstanceOf(Cyan::class, $cyan2);
         $this->assertNotSame($cyan, $cyan2);
@@ -49,9 +49,9 @@ class CyanTest extends TestCase
 
     public function testSubtract()
     {
-        $cyan = new Cyan(50);
+        $cyan = Cyan::at(50);
 
-        $cyan2 = $cyan->subtract(new Cyan(25));
+        $cyan2 = $cyan->subtract(Cyan::at(25));
 
         $this->assertInstanceOf(Cyan::class, $cyan2);
         $this->assertNotSame($cyan, $cyan2);
@@ -61,21 +61,21 @@ class CyanTest extends TestCase
 
     public function testEquals()
     {
-        $this->assertTrue((new Cyan(50))->equals(new Cyan(50)));
-        $this->assertFalse((new Cyan(100))->equals(new Cyan(50)));
+        $this->assertTrue(Cyan::at(50)->equals(Cyan::at(50)));
+        $this->assertFalse(Cyan::at(100)->equals(Cyan::at(50)));
     }
 
     public function testAtMaximum()
     {
-        $this->assertTrue((new Cyan(100))->atMaximum());
-        $this->assertFalse((new Cyan(0))->atMaximum());
-        $this->assertFalse((new Cyan(50))->atMaximum());
+        $this->assertTrue(Cyan::at(100)->atMaximum());
+        $this->assertFalse(Cyan::at(0)->atMaximum());
+        $this->assertFalse(Cyan::at(50)->atMaximum());
     }
 
     public function testAtMinimum()
     {
-        $this->assertFalse((new Cyan(100))->atMinimum());
-        $this->assertTrue((new Cyan(0))->atMinimum());
-        $this->assertFalse((new Cyan(50))->atMinimum());
+        $this->assertFalse(Cyan::at(100)->atMinimum());
+        $this->assertTrue(Cyan::at(0)->atMinimum());
+        $this->assertFalse(Cyan::at(50)->atMinimum());
     }
 }