+++ CHANGELOG.md
@@ -6,6 +6,21 @@
 
 - Requires PHP `8.4`
 - Named constructors `::of()`, `::fromHexadecimal()` and `::fromIntensity()` now return instances of `Innmind\Immutable\Attempt`
+- `Innmind\Colour\Alpha` constructor is now private, use `::of()` instead
+- `Innmind\Colour\Back` constructor is now private, use `::of()` or `::at()` instead
+- `Innmind\Colour\Blue` constructor is now private, use `::of()` or `::at()` instead
+- `Innmind\Colour\Cyan` constructor is now private, use `::of()` or `::at()` instead
+- `Innmind\Colour\Green` constructor is now private, use `::of()` or `::at()` instead
+- `Innmind\Colour\Hue` constructor is now private, use `::of()` or `::at()` instead
+- `Innmind\Colour\Intensity` constructor is now private, use `::of()` or `::at()` instead
+- `Innmind\Colour\Lightness` constructor is now private, use `::of()` or `::at()` instead
+- `Innmind\Colour\Magenta` constructor is now private, use `::of()` or `::at()` instead
+- `Innmind\Colour\Red` constructor is now private, use `::of()` or `::at()` instead
+- `Innmind\Colour\Saturation` constructor is now private, use `::of()` or `::at()` instead
+- `Innmind\Colour\Yellow` constructor is now private, use `::of()` or `::at()` instead
+- `Innmind\Colour\CMYKA` constructor is now private, use `::from()` instead
+- `Innmind\Colour\HSLA` constructor is now private, use `::from()` instead
+- `Innmind\Colour\RGBA` constructor is now private, use `::from()` instead
 
 ### Fixed